checkpoint work

+++ glop/tests/main_test.py
@@ -1,4 +1,4 @@
-# Copyright 2014 Google Inc. All rights reserved.
+# Copyright 2014 Dirk Pranke. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -14,8 +14,8 @@
 
 import unittest
 
-from glop.host import Host
 from glop.fakes.host_fake import FakeHost
+from glop.host import Host
 from glop.main import main, VERSION
 
 
@@ -25,12 +25,12 @@ SIMPLE_GRAMMAR = "grammar = anything*:as end -> ''.join(as) ,"
 class CheckMixin(object):
     def _write_files(self, host, files):
         for path, contents in list(files.items()):
-            host.write(path, contents)
+            host.write_text_file(path, contents)
 
     def _read_files(self, host, tmpdir):
         out_files = {}
         for f in host.files_under(tmpdir):
-            out_files[f] = host.read(tmpdir, f)
+            out_files[f] = host.read_text_file(tmpdir, f)
         return out_files
 
     def assert_files(self, expected_files, actual_files):
@@ -42,9 +42,11 @@ class CheckMixin(object):
         host = self._host()
         try:
             tmpdir = host.mkdtemp()
-            fname = host.join(tmpdir, 'input.txt')
-            host.write(fname, input_txt)
-            args = ['-i', '-e', grammar, fname]
+            input_path = host.join(tmpdir, 'input.txt')
+            grammar_path = host.join(tmpdir, 'grammar.g')
+            host.write_text_file(input_path, input_txt)
+            host.write_text_file(grammar_path, grammar)
+            args = ['--interpret', '-i', input_path, grammar_path]
             self._call(host, args, None, returncode, out, err)
         finally:
             host.rmtree(tmpdir)
@@ -52,6 +54,7 @@ class CheckMixin(object):
     def check_cmd(self, args, stdin=None, files=None,
                   returncode=None, out=None, err=None, output_files=None):
         host = self._host()
+        orig_wd, tmpdir = None, None
         try:
             orig_wd = host.getcwd()
             tmpdir = host.mkdtemp()
@@ -62,8 +65,10 @@ class CheckMixin(object):
             actual_ret, actual_out, actual_err = rv
             actual_output_files = self._read_files(host, host.getcwd())
         finally:
-            host.rmtree(tmpdir)
-            host.chdir(orig_wd)
+            if tmpdir:
+              host.rmtree(tmpdir)
+            if orig_wd:
+              host.chdir(orig_wd)
 
         if output_files:
             self.assert_files(output_files, actual_output_files)
@@ -71,18 +76,14 @@ class CheckMixin(object):
 
 
 class UnitTestMixin(object):
-    use_compiled_grammar_parser = False
-
     def _host(self):
         return FakeHost()
 
     def _call(self, host, args, stdin=None,
               returncode=None, out=None, err=None):
         if stdin is not None:
-            host.stdin.write(stdin)
+            host.stdin.write(unicode(stdin))
             host.stdin.seek(0)
-        if self.use_compiled_grammar_parser:
-            args = ['--use-compiled-grammar-parser'] + args
         actual_ret = main(host, args)
         actual_out = host.stdout.getvalue()
         actual_err = host.stderr.getvalue()
@@ -123,11 +124,11 @@ class UnitTestMain(UnitTestMixin, CheckMixin, unittest.TestCase):
         def raise_ctrl_c(*_comps):
             raise KeyboardInterrupt
 
-        host.read = raise_ctrl_c
-        host.write('simple.g', SIMPLE_GRAMMAR)
+        host.read_text_file = raise_ctrl_c
+        host.write_text_file('simple.g', SIMPLE_GRAMMAR)
 
-        self._call(host, ['-i', 'simple.g'], returncode=130,
-                   out='', err='Interrupted, exiting ..\n')
+        self._call(host, ['simple.g'], returncode=130,
+                   out='', err='Interrupted, exiting ...\n')
 
 
 class TestMain(UnitTestMixin, CheckMixin, unittest.TestCase):
@@ -138,41 +139,42 @@ class TestMain(UnitTestMixin, CheckMixin, unittest.TestCase):
         }
         out_files = files.copy()
         out_files['output.txt'] = 'hello, world\n'
-        self.check_cmd(['-i', '-o', 'output.txt', 'simple.g', 'input.txt'],
+        self.check_cmd(['--interpret', '-i', 'input.txt', '-o', 'output.txt',
+                        'simple.g'],
                        files=files, returncode=0, out='', err='',
                        output_files=out_files)
 
     def test_no_grammar(self):
-        self.check_cmd([], returncode=1,
-                       err='Must specify a grammar file or a string with -e.\n')
+        self.check_cmd([], returncode=2,
+                       err='Must provide a grammar file.\n')
 
     def test_grammar_file_not_found(self):
-        self.check_cmd(['-i', 'missing.g'], returncode=1,
-                       err='grammar file "missing.g" not found\n')
+        self.check_cmd(['missing.g'], returncode=1,
+                       err='Error: no such file: "missing.g"\n')
 
     def test_input_on_stdin(self):
-        self.check_cmd(['-i', '-e', SIMPLE_GRAMMAR], stdin="hello, world\n",
-                       returncode=0, out="hello, world\n", err='')
+        files = {
+            'simple.g': SIMPLE_GRAMMAR,
+        }
+        self.check_cmd(['--interpret', 'simple.g'], stdin="hello, world\n",
+                       files=files, returncode=0, out="hello, world\n", err='')
 
     def test_print_grammar_to_out(self):
         files = {
             'simple.g': SIMPLE_GRAMMAR,
         }
         out_files = files.copy()
-        self.check_cmd(['-p', 'simple.g'], files=files,
+        self.check_cmd(['--pretty-print', 'simple.g'], files=files,
                        returncode=0,
                        out="grammar = anything*:as end -> ''.join(as),\n",
                        output_files=out_files)
 
-    def test_print_bad_grammar(self):
-        self.check_cmd(['-p', '-e', 'grammar ='],
-                       returncode=1, out='',
-                       err='-e:1:2 expecting the end\n')
-
     def test_parse_bad_grammar(self):
-        self.check_cmd(['-i', '-e', 'grammar ='],
-                       returncode=1, out='',
-                       err='-e:1:2 expecting the end\n')
+        files = {
+            'bad.g': 'grammar =',
+        }
+        self.check_cmd(['bad.g'], files=files,
+                       returncode=1, out='', err=None)
 
     def test_version(self):
         self.check_cmd(['-V'], returncode=0, out=(VERSION + '\n'),
@@ -240,10 +242,6 @@ class TestInterpreter(UnitTestMixin, CheckMixin, unittest.TestCase):
     def test_py_getitem(self):
         self.check_match("grammar = end -> 'bar'[1] ,", '',
                          returncode=0, out='a')
-
-
-class TestCompiledParser(TestInterpreter):
-    use_compiled_grammar_parser = True
 
 
 if __name__ == '__main__':
